Support console: combine user creation code
Before this commit, there were two code paths, one for inviting a user to
a school and another for inviting a user to a responsible body. The two paths
had a lot of similarities and small inconsistencies. These have been rationalised
into two controller actions on Support::UsersController.
