--- app/views/support/schools/users/new.html.erb
@@ -1,25 +1,0 @@
-<%- title = t('page_titles.new_school_user') %>
-<%- content_for :title, title_with_error_prefix(title, @user.errors.any?) %>
-<%- content_for :before_content do %>
-  <%= breadcrumbs([
-    { 'Support home' => support_home_path },
-    { t('page_titles.support_responsible_bodies') => support_responsible_bodies_path },
-    { @school.responsible_body.name => support_responsible_body_path(@school.responsible_body) },
-    { @school.name => support_school_path(urn: @school.urn) },
-    title,
-  ]) %>
-<% end %>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-xl">
-      <%= title %>
-    </h1>
-
-    <%= form_for @user, url: support_school_users_path(@school) do |f| %>
-      <%= f.govuk_error_summary %>
-      <%= render partial: 'shared/school_user_form', locals: { form: f } %>
-      <%= f.govuk_submit 'Send invite' %>
-    <%- end %>
-  </div>
-</div>
